+++ LookupAnything/Framework/Lookups/Items/ItemSubject.cs
@@ -19,6 +19,7 @@ using StardewValley.GameData.FishPond;
 using StardewValley.GameData.Movies;
 using StardewValley.Locations;
 using StardewValley.Objects;
+using StardewValley.TerrainFeatures;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
@@ -38,6 +39,9 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         /// <summary>The crop which will drop the item (if applicable).</summary>
         private readonly Crop FromCrop;
 
+        /// <summary>The dirt containing the crop (if applicable).</summary>
+        private readonly HoeDirt FromDirt;
+
         /// <summary>The crop grown by this seed item (if applicable).</summary>
         private readonly Crop SeedForCrop;
 
@@ -54,7 +58,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         private readonly bool HighlightUnrevealedGiftTastes;
 
         /// <summary>Get a lookup subject for a crop.</summary>
-        private readonly Func<Crop, ObjectContext, ISubject> GetCropSubject;
+        private readonly Func<Crop, ObjectContext, HoeDirt, ISubject> GetCropSubject;
 
 
         /*********
@@ -69,14 +73,17 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         /// <param name="knownQuality">Whether the item quality is known. This is <c>true</c> for an inventory item, <c>false</c> for a map object.</param>
         /// <param name="getCropSubject">Get a lookup subject for a crop.</param>
         /// <param name="fromCrop">The crop associated with the item (if applicable).</param>
-        public ItemSubject(GameHelper gameHelper, bool progressionMode, bool highlightUnrevealedGiftTastes, Item item, ObjectContext context, bool knownQuality, Func<Crop, ObjectContext, ISubject> getCropSubject, Crop fromCrop = null)
+        /// <param name="fromDirt">The dirt containing the crop (if applicable).</param>
+        public ItemSubject(GameHelper gameHelper, bool progressionMode, bool highlightUnrevealedGiftTastes, Item item, ObjectContext context, bool knownQuality, Func<Crop, ObjectContext, HoeDirt, ISubject> getCropSubject, Crop fromCrop = null, HoeDirt fromDirt = null)
             : base(gameHelper)
         {
             this.ProgressionMode = progressionMode;
             this.HighlightUnrevealedGiftTastes = highlightUnrevealedGiftTastes;
             this.Target = item;
             this.DisplayItem = this.GetMenuItem(item);
-            this.FromCrop = fromCrop;
+            this.FromCrop = fromCrop ?? fromDirt?.crop;
+            this.FromDirt = fromDirt;
+
             if ((item as SObject)?.Type == "Seeds" && fromCrop == null) // fromCrop == null to exclude planted coffee beans
                 this.SeedForCrop = new Crop(item.ParentSheetIndex, 0, 0);
             this.Context = context;
@@ -119,7 +126,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
             }
 
             // crop fields
-            foreach (ICustomField field in this.GetCropFields(this.FromCrop ?? this.SeedForCrop, isSeed))
+            foreach (ICustomField field in this.GetCropFields(this.FromDirt, this.FromCrop ?? this.SeedForCrop, isSeed))
                 yield return field;
 
             // indoor pot crop
@@ -129,7 +136,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                 if (potCrop != null)
                 {
                     Item drop = this.GameHelper.GetObjectBySpriteIndex(potCrop.indexOfHarvest.Value);
-                    yield return new LinkField(I18n.Item_Contents(), drop.DisplayName, () => this.GetCropSubject(potCrop, ObjectContext.World));
+                    yield return new LinkField(I18n.Item_Contents(), drop.DisplayName, () => this.GetCropSubject(potCrop, ObjectContext.World, pot.hoeDirt.Value));
                 }
             }
 
@@ -310,7 +317,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
             if (seeAlsoCrop)
             {
                 Item drop = this.GameHelper.GetObjectBySpriteIndex(this.SeedForCrop.indexOfHarvest.Value);
-                yield return new LinkField(I18n.Item_SeeAlso(), drop.DisplayName, () => this.GetCropSubject(this.SeedForCrop, ObjectContext.Inventory));
+                yield return new LinkField(I18n.Item_SeeAlso(), drop.DisplayName, () => this.GetCropSubject(this.SeedForCrop, ObjectContext.Inventory, null));
             }
         }
 
@@ -401,14 +408,16 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
         }
 
         /// <summary>Get the custom fields for a crop.</summary>
+        /// <param name="dirt">The dirt the crop is planted in, if applicable.</param>
         /// <param name="crop">The crop to represent.</param>
         /// <param name="isSeed">Whether the crop being displayed is for an unplanted seed.</param>
-        private IEnumerable<ICustomField> GetCropFields(Crop crop, bool isSeed)
+        private IEnumerable<ICustomField> GetCropFields(HoeDirt dirt, Crop crop, bool isSeed)
         {
             if (crop == null)
                 yield break;
 
             var data = new CropDataParser(crop, isPlanted: !isSeed);
+            bool isForage = crop.whichForageCrop.Value > 0;
 
             // add next-harvest field
             if (!isSeed)
@@ -429,7 +438,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
             }
 
             // crop summary
-            if (crop.whichForageCrop.Value <= 0)
+            if (!isForage)
             {
                 List<string> summary = new List<string>();
 
@@ -457,6 +466,33 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
 
                 // generate field
                 yield return new GenericField(I18n.Crop_Summary(), "-" + string.Join($"{Environment.NewLine}-", summary));
+            }
+
+            // dirt water/fertilizer state
+            if (dirt != null && !isForage)
+            {
+                // watered
+                yield return new GenericField(I18n.Crop_Watered(), this.Stringify(dirt.state.Value == HoeDirt.watered));
+
+                // fertilizer
+                yield return new GenericField(I18n.Crop_Fertilized(), dirt.fertilizer.Value switch
+                {
+                    HoeDirt.noFertilizer => this.Stringify(false),
+
+                    HoeDirt.speedGro => GameI18n.GetObjectName(465), // Speed-Gro
+                    HoeDirt.superSpeedGro => GameI18n.GetObjectName(466), // Deluxe Speed-Gro
+                    HoeDirt.hyperSpeedGro => GameI18n.GetObjectName(918), // Hyper Speed-Gro
+
+                    HoeDirt.fertilizerLowQuality => GameI18n.GetObjectName(368), // Basic Fertilizer
+                    HoeDirt.fertilizerHighQuality => GameI18n.GetObjectName(919), // Deluxe Fertilizer
+                    HoeDirt.fertilizerDeluxeQuality => GameI18n.GetObjectName(369), // Quality Fertilizer
+
+                    HoeDirt.waterRetentionSoil => GameI18n.GetObjectName(370), // Basic Retaining Soil
+                    HoeDirt.waterRetentionSoilQuality => GameI18n.GetObjectName(371), // Quality Retaining Soil
+                    HoeDirt.waterRetentionSoilDeluxe => GameI18n.GetObjectName(920), // Deluxe Retaining Soil
+
+                    _ => I18n.Generic_Unknown()
+                });
             }
         }