+++ SmallBeachFarm/Framework/FishType.cs
@@ -7,6 +7,9 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
         River,
 
         /// <summary>Fish that can be caught in the ocean.</summary>
-        Ocean
+        Ocean,
+
+        /// <summary>Fish from a non-open-water position, like a fish pond.</summary>
+        Default
     }
 }