+++ Automate/release-notes.md
@@ -1,6 +1,13 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.19.1
+Released 04 January 2021.
+
+* When you join a multiplayer game, Automate shows a warning that it needs to be installed by the main player. That message is now clearer and only shown if they don't have it installed.
+* Fixed slime egg incubators ignoring tiger slime eggs.
+* Fixed deconstructor taking multiple inputs from the chest, but only producing output for the last one.
+
 ## 1.19
 Released 21 December 2020.