+++ DataLayers/release-notes.md
@@ -1,4 +1,10 @@
 # Release notes
+## 1.13.2
+Released 04 January 2021.
+
+* Fixed sprinkler layer not showing range for older sprinkler mods in Stardew Valley 1.5.
+* Improved translations. Thanks to elCrimar (updated Spanish) and norges (updated German)!
+
 ## 1.13.1
 Released 21 December 2020.