+++ FastAnimations/release-notes.md
@@ -1,6 +1,11 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.9.4
+Released 04 January 2021.
+
+* Updated for Tractor Mod 4.13.
+
 ## 1.9.3
 Released 21 December 2020.