+++ SmallBeachFarm/release-notes.md
@@ -1,6 +1,12 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.8.1
+Released 04 January 2021.
+
+* Fixed issues when moving greenhouse or default shipping bin in Stardew Valley 1.5.
+* Fixed fishing in fish ponds on the small beach farm.
+
 ## 1.8
 Released 12 September 2020.